Move icon test data into test folder
