+++ packages/terra-core-docs/src/terra-dev-site/doc/icon/IconCollectionTable/index.js
@@ -1,5 +1,5 @@
 import IconCollectionTable from './IconCollectionTable';
-import iconData from './iconData';
+import iconData from '../../../test/icon-library-table/iconData';
 
 export {IconCollectionTable, iconData};
 export default IconCollectionTable;
